change longs to unsigned long long

+++ iosapi/iosapi/PNEvent.h
@@ -4,17 +4,17 @@
 @interface PNEvent : NSObject <NSCoding> {
     PNEventType _eventType;
     NSTimeInterval  _eventTime;
-    long _applicationId;
+    unsigned long long _applicationId;
     NSString * _userId;
 }
 
 @property (nonatomic, retain) NSString * sessionId;
 @property (nonatomic, assign) PNEventType eventType;
 @property (nonatomic, assign) NSTimeInterval eventTime;
-@property (nonatomic, assign) long applicationId;
+@property (nonatomic, assign) unsigned long long applicationId;
 @property (nonatomic, retain) NSString *userId;
 
-- (id) init: (PNEventType)eventType applicationId:(long) applicationId userId:(NSString *)userId;
+- (id) init: (PNEventType)eventType applicationId:(unsigned long long) applicationId userId:(NSString *)userId;
 - (NSString *) description;
 - (NSString *) addOptionalParam:(NSString *)url name:(NSString *)name value:(NSString *)value;
 - (NSString *) toQueryString;
